[JABERA-13] Move userEntity to auth package
