+++ src/main/java/com/the/dark/side/crew/fejsbuk/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.the.dark.side.crew.fejsbuk.repository;
 
-import com.the.dark.side.crew.fejsbuk.model.UserEntity;
+import com.the.dark.side.crew.fejsbuk.auth.domain.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
