+++ src/main/java/com/the/dark/side/crew/fejsbuk/service/UserService.java
@@ -1,6 +1,6 @@
 package com.the.dark.side.crew.fejsbuk.service;
 
-import com.the.dark.side.crew.fejsbuk.model.UserEntity;
+import com.the.dark.side.crew.fejsbuk.auth.domain.entity.UserEntity;
 import java.util.Optional;
 
 public interface UserService {
