+++ src/test/java/com/the/dark/side/crew/fejsbuk/service/impl/PostServiceTest.java
@@ -2,7 +2,7 @@ package com.the.dark.side.crew.fejsbuk.service.impl;
 
 import com.the.dark.side.crew.fejsbuk.mapper.PostMapper;
 import com.the.dark.side.crew.fejsbuk.model.PostEntity;
-import com.the.dark.side.crew.fejsbuk.model.UserEntity;
+import com.the.dark.side.crew.fejsbuk.auth.domain.entity.UserEntity;
 import com.the.dark.side.crew.fejsbuk.model.dto.PostDto;
 import com.the.dark.side.crew.fejsbuk.repository.PostRepository;
 import org.junit.jupiter.api.BeforeEach;
